# backend-go/api/openapi.json
{"components":{"schemas":{"AnnotationCreate":{"properties":{"body":{"minLength":1,"title":"Body","type":"string"},"color":{"default":"#2563eb","pattern":"^#[0-9a-fA-F]{6}$","title":"Color","type":"string"},"custom_data":{"additionalProperties":true,"title":"Custom Data","type":"object"},"display_mode":{"default":"card","maxLength":60,"minLength":1,"title":"Display Mode","type":"string"},"duration_seconds":{"default":6.0,"exclusiveMinimum":0.0,"maximum":3600.0,"title":"Duration Seconds","type":"number"},"interactive":{"default":true,"title":"Interactive","type":"boolean"},"kind":{"default":"note","maxLength":60,"minLength":1,"title":"Kind","type":"string"},"position_x":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Position X"},"position_y":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Position Y"},"region_height":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Region Height"},"region_width":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Region Width"},"region_x":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Region X"},"region_y":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Region Y"},"shape":{"default":"marker","maxLength":60,"minLength":1,"title":"Shape","type":"string"},"timestamp_seconds":{"minimum":0.0,"title":"Timestamp Seconds","type":"number"},"title":{"maxLength":200,"minLength":1,"title":"Title","type":"string"}},"required":["timestamp_seconds","title","body"],"title":"AnnotationCreate","type":"object"},"AnnotationRead":{"properties":{"body":{"title":"Body","type":"string"},"color":{"title":"Color","type":"string"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"custom_data":{"additionalProperties":true,"title":"Custom Data","type":"object"},"display_mode":{"title":"Display Mode","type":"string"},"duration_seconds":{"title":"Duration Seconds","type":"number"},"id":{"format":"uuid","title":"Id","type":"string"},"interactive":{"title":"Interactive","type":"boolean"},"kind":{"title":"Kind","type":"string"},"position_x":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Position X"},"position_y":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Position Y"},"region_height":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Region Height"},"region_width":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Region Width"},"region_x":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Region X"},"region_y":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Region Y"},"shape":{"title":"Shape","type":"string"},"timestamp_seconds":{"title":"Timestamp Seconds","type":"number"},"title":{"title":"Title","type":"string"},"updated_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Updated At"},"video_id":{"format":"uuid","title":"Video Id","type":"string"}},"required":["id","video_id","timestamp_seconds","duration_seconds","position_x","position_y","region_x","region_y","region_width","region_height","shape","display_mode","interactive","title","body","kind","color","custom_data","created_at","updated_at"],"title":"AnnotationRead","type":"object"},"AnnotationUpdate":{"properties":{"body":{"anyOf":[{"minLength":1,"type":"string"},{"type":"null"}],"title":"Body"},"color":{"anyOf":[{"pattern":"^#[0-9a-fA-F]{6}$","type":"string"},{"type":"null"}],"title":"Color"},"custom_data":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Custom Data"},"display_mode":{"anyOf":[{"maxLength":60,"minLength":1,"type":"string"},{"type":"null"}],"title":"Display Mode"},"duration_seconds":{"anyOf":[{"exclusiveMinimum":0.0,"maximum":3600.0,"type":"number"},{"type":"null"}],"title":"Duration Seconds"},"interactive":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Interactive"},"kind":{"anyOf":[{"maxLength":60,"minLength":1,"type":"string"},{"type":"null"}],"title":"Kind"},"position_x":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Position X"},"position_y":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Position Y"},"region_height":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Region Height"},"region_width":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Region Width"},"region_x":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Region X"},"region_y":{"anyOf":[{"maximum":1.0,"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Region Y"},"shape":{"anyOf":[{"maxLength":60,"minLength":1,"type":"string"},{"type":"null"}],"title":"Shape"},"timestamp_seconds":{"anyOf":[{"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Timestamp Seconds"},"title":{"anyOf":[{"maxLength":200,"minLength":1,"type":"string"},{"type":"null"}],"title":"Title"}},"title":"AnnotationUpdate","type":"object"},"GroupCreate":{"properties":{"name":{"maxLength":120,"minLength":1,"title":"Name","type":"string"}},"required":["name"],"title":"GroupCreate","type":"object"},"GroupRead":{"properties":{"created_at":{"format":"date-time","title":"Created At","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"name":{"title":"Name","type":"string"}},"required":["id","name","created_at"],"title":"GroupRead","type":"object"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"PaginatedResponse_VideoRead_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/VideoRead"},"title":"Items","type":"array"},"page":{"title":"Page","type":"integer"},"page_size":{"title":"Page Size","type":"integer"},"total":{"title":"Total","type":"integer"}},"required":["items","total","page","page_size"],"title":"PaginatedResponse[VideoRead]","type":"object"},"TokenRead":{"properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"default":"bearer","title":"Token Type","type":"string"},"user":{"$ref":"#/components/schemas/UserRead"}},"required":["access_token","user"],"title":"TokenRead","type":"object"},"UploadAbortRequest":{"properties":{"object_key":{"minLength":1,"title":"Object Key","type":"string"},"upload_id":{"minLength":1,"title":"Upload Id","type":"string"}},"required":["object_key","upload_id"],"title":"UploadAbortRequest","type":"object"},"UploadCompleteRequest":{"properties":{"content_type":{"maxLength":255,"minLength":1,"title":"Content Type","type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"filename":{"maxLength":512,"minLength":1,"title":"Filename","type":"string"},"object_key":{"minLength":1,"title":"Object Key","type":"string"},"parts":{"items":{"$ref":"#/components/schemas/UploadPartCompleted"},"minItems":1,"title":"Parts","type":"array"},"size_bytes":{"exclusiveMinimum":0.0,"title":"Size Bytes","type":"integer"},"title":{"maxLength":200,"minLength":1,"title":"Title","type":"string"},"upload_id":{"minLength":1,"title":"Upload Id","type":"string"}},"required":["object_key","upload_id","title","filename","content_type","size_bytes","parts"],"title":"UploadCompleteRequest","type":"object"},"UploadInitRequest":{"properties":{"content_type":{"maxLength":255,"minLength":1,"title":"Content Type","type":"string"},"filename":{"maxLength":512,"minLength":1,"title":"Filename","type":"string"},"size_bytes":{"exclusiveMinimum":0.0,"title":"Size Bytes","type":"integer"}},"required":["filename","content_type","size_bytes"],"title":"UploadInitRequest","type":"object"},"UploadInitResponse":{"properties":{"concurrency":{"title":"Concurrency","type":"integer"},"expires_in":{"title":"Expires In","type":"integer"},"object_key":{"title":"Object Key","type":"string"},"part_size":{"title":"Part Size","type":"integer"},"parts":{"items":{"$ref":"#/components/schemas/UploadPartUrl"},"title":"Parts","type":"array"},"upload_id":{"title":"Upload Id","type":"string"}},"required":["object_key","upload_id","part_size","parts","expires_in","concurrency"],"title":"UploadInitResponse","type":"object"},"UploadPartCompleted":{"properties":{"etag":{"maxLength":512,"minLength":1,"title":"Etag","type":"string"},"part_number":{"maximum":10000.0,"minimum":1.0,"title":"Part Number","type":"integer"}},"required":["part_number","etag"],"title":"UploadPartCompleted","type":"object"},"UploadPartUrl":{"properties":{"part_number":{"title":"Part Number","type":"integer"},"url":{"title":"Url","type":"string"}},"required":["part_number","url"],"title":"UploadPartUrl","type":"object"},"UserCreate":{"properties":{"group_id":{"format":"uuid","title":"Group Id","type":"string"},"password":{"maxLength":256,"minLength":8,"title":"Password","type":"string"},"username":{"maxLength":80,"minLength":3,"title":"Username","type":"string"}},"required":["username","password","group_id"],"title":"UserCreate","type":"object"},"UserLogin":{"properties":{"password":{"title":"Password","type":"string"},"username":{"title":"Username","type":"string"}},"required":["username","password"],"title":"UserLogin","type":"object"},"UserRead":{"properties":{"created_at":{"format":"date-time","title":"Created At","type":"string"},"group_id":{"format":"uuid","title":"Group Id","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"is_admin":{"title":"Is Admin","type":"boolean"},"username":{"title":"Username","type":"string"}},"required":["id","group_id","username","is_admin","created_at"],"title":"UserRead","type":"object"},"ValidationError":{"properties":{"ctx":{"title":"Context","type":"object"},"input":{"title":"Input"},"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"},"VideoRead":{"properties":{"content_type":{"title":"Content Type","type":"string"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"group_id":{"format":"uuid","title":"Group Id","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"original_filename":{"title":"Original Filename","type":"string"},"playback_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Playback Url"},"size_bytes":{"title":"Size Bytes","type":"integer"},"title":{"title":"Title","type":"string"},"updated_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Updated At"}},"required":["id","group_id","title","description","original_filename","content_type","size_bytes","created_at","updated_at"],"title":"VideoRead","type":"object"},"VideoUpdate":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"title":{"anyOf":[{"maxLength":200,"minLength":1,"type":"string"},{"type":"null"}],"title":"Title"}},"title":"VideoUpdate","type":"object"}},"securitySchemes":{"HTTPBearer":{"scheme":"bearer","type":"http"}}},"info":{"title":"VideoInsight API","version":"0.1.0"},"openapi":"3.1.0","paths":{"/api/annotations/{annotation_id}":{"delete":{"operationId":"delete_annotation_api_annotations__annotation_id__delete","parameters":[{"in":"path","name":"annotation_id","required":true,"schema":{"format":"uuid","title":"Annotation Id","type":"string"}}],"responses":{"204":{"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Delete Annotation","tags":["annotations"]},"patch":{"operationId":"update_annotation_api_annotations__annotation_id__patch","parameters":[{"in":"path","name":"annotation_id","required":true,"schema":{"format":"uuid","title":"Annotation Id","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnnotationUpdate"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnnotationRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Update Annotation","tags":["annotations"]}},"/api/auth/login":{"post":{"operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Login","tags":["auth"]}},"/api/auth/me":{"get":{"operationId":"me_api_auth_me_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}},"description":"Successful Response"}},"security":[{"HTTPBearer":[]}],"summary":"Me","tags":["auth"]}},"/api/auth/users":{"post":{"operationId":"create_user_api_auth_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Create User","tags":["auth"]}},"/api/groups":{"get":{"operationId":"list_groups_api_groups_get","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GroupRead"},"title":"Response List Groups Api Groups Get","type":"array"}}},"description":"Successful Response"}},"security":[{"HTTPBearer":[]}],"summary":"List Groups","tags":["groups"]},"post":{"operationId":"create_group_api_groups_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreate"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Create Group","tags":["groups"]}},"/api/health":{"get":{"operationId":"health_check_api_health_get","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Health Check","tags":["health"]}},"/api/videos":{"get":{"operationId":"list_videos_api_videos_get","parameters":[{"in":"query","name":"page","required":false,"schema":{"default":1,"minimum":1,"title":"Page","type":"integer"}},{"in":"query","name":"page_size","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Page Size","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_VideoRead_"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"List Videos","tags":["videos"]},"post":{"operationId":"create_video_api_videos_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadCompleteRequest"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Create Video","tags":["videos"]}},"/api/videos/uploads":{"post":{"operationId":"init_upload_api_videos_uploads_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadInitRequest"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadInitResponse"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Init Upload","tags":["videos"]}},"/api/videos/uploads/abort":{"post":{"operationId":"abort_upload_api_videos_uploads_abort_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadAbortRequest"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Abort Upload","tags":["videos"]}},"/api/videos/{video_id}":{"delete":{"operationId":"delete_video_api_videos__video_id__delete","parameters":[{"in":"path","name":"video_id","required":true,"schema":{"format":"uuid","title":"Video Id","type":"string"}}],"responses":{"204":{"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Delete Video","tags":["videos"]},"get":{"operationId":"get_video_api_videos__video_id__get","parameters":[{"in":"path","name":"video_id","required":true,"schema":{"format":"uuid","title":"Video Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Get Video","tags":["videos"]},"patch":{"operationId":"update_video_api_videos__video_id__patch","parameters":[{"in":"path","name":"video_id","required":true,"schema":{"format":"uuid","title":"Video Id","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoUpdate"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Update Video","tags":["videos"]}},"/api/videos/{video_id}/annotations":{"get":{"operationId":"list_annotations_api_videos__video_id__annotations_get","parameters":[{"in":"path","name":"video_id","required":true,"schema":{"format":"uuid","title":"Video Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AnnotationRead"},"title":"Response List Annotations Api Videos  Video Id  Annotations Get","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"List Annotations","tags":["annotations"]},"post":{"operationId":"create_annotation_api_videos__video_id__annotations_post","parameters":[{"in":"path","name":"video_id","required":true,"schema":{"format":"uuid","title":"Video Id","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnnotationCreate"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnnotationRead"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"security":[{"HTTPBearer":[]}],"summary":"Create Annotation","tags":["annotations"]}}}}